Added tests to check error messages
